wiring up fqs query
