+++ src/fhir/patient-helper.ts
@@ -7,7 +7,9 @@ import { errorResponse } from "@/utils/errors";
 import { pickBy } from "@/utils/nodash";
 import { hasNumber } from "@/utils/types";
 import { PatientGraphqlResponse, patientsForBuilderQuery } from "@/services/fqs/queries/patients";
-import { createGraphqlClient, fqsRequest } from "..";
+import { createGraphqlClient, fqsRequest, useQueryWithCTW } from "..";
+import { QUERY_KEY_PATIENTS_LIST } from "@/utils/query-keys";
+import { first } from "lodash";
 
 export async function getBuilderFhirPatient(
   requestContext: CTWRequestContext,
@@ -37,6 +39,14 @@ export async function getBuilderFhirPatient(
   }
 
   return new PatientModel(patients[0], getIncludedResources(bundle));
+}
+
+export function usePatientsList(pageSize: number, pageOffset: number, searchNameValue?: string) {
+  return useQueryWithCTW(
+    QUERY_KEY_PATIENTS_LIST,
+    [pageSize, pageOffset, searchNameValue],
+    getBuilderPatientsList
+  );
 }
 
 type GetPatientsTableResults = {
@@ -71,11 +81,12 @@ export async function getBuilderPatientListWithSearch(
   }
 }
 
-export async function getPatientsForBuilder(
+export async function getBuilderPatientsList(
   requestContext: CTWRequestContext,
-  first: number,
-  cursor?: string
-) {
+  paginationOptions: (number | string | undefined)[] = []
+): Promise<PatientModel[]> {
+  const [pageSize, pageOffset, searchValue] = paginationOptions;
+
   try {
     const graphClient = createGraphqlClient(requestContext);
     const { data } = await fqsRequest<PatientGraphqlResponse>(
@@ -83,8 +94,8 @@ export async function getPatientsForBuilder(
       patientsForBuilderQuery,
       {
         builderID: requestContext.builderId,
-        cursor: cursor ?? "",
-        first: first,
+        cursor: "",
+        first: pageSize,
         sort: {
           lastUpdated: "DESC",
         },
@@ -93,38 +104,6 @@ export async function getPatientsForBuilder(
     return data.PatientConnection.edges.map((x) => new PatientModel(x.node));
   } catch (e) {
     throw new Error(`Failed fetching patients: ${e}`);
-  }
-}
-
-export async function getBuilderPatientsList(
-  requestContext: CTWRequestContext,
-  paginationOptions: (number | string | undefined)[] = []
-): Promise<GetPatientsTableResults> {
-  const [pageSize, pageOffset, searchValue] = paginationOptions;
-  const offset = parseInt(`${pageOffset ?? "0"}`, 10) * parseInt(`${pageSize ?? "1"}`, 10);
-
-  const searchParams = pickBy({
-    _count: pageSize,
-    _total: "accurate",
-    _offset: offset,
-    _sort: "family",
-    ...(hasNumber(searchValue) ? { identifier: searchValue } : { name: searchValue }),
-  }) as SearchParams;
-
-  try {
-    const { total, resources } = await searchBuilderRecords(
-      "Patient",
-      requestContext,
-      searchParams
-    );
-
-    return {
-      searchParams,
-      patients: resources.map((patient) => new PatientModel(patient)),
-      total,
-    };
-  } catch (e) {
-    throw errorResponse("Failed fetching patients", e);
   }
 }
